Use nbtree knn search with ScalarArrayOp and backwardscans optimizations.
When ordering by distance with SAOP qual, there are three possible cases:
- case A: ordering argument is less or equal to the min array element;
- case B: ordering argument is greater or equal to the max element;
- case C: ordering argument is within the array bounds.

Fully use optimizations from 5bf748b86 and 1bd4bc85c in case of
knn ordering when the ordering key argument is to the left (A)
or right of the array (B).

Add abilty to spin scan direction globally for this.

Exclude forced re-sorting arrays by distance in most cases.

Add abiity for merge arrays sorted by distance.

Add in-place deduplication of such an arrays.

Add some simple tests for different locations of the distance
start point and for arrays deduplication and merging.

For case (C) temporary leave the existing behaviour with
index descent per each array element although partial use
of these optimizations is also possible. Even in the worst case,
when the ordering argument falls in the middle of array, the
number of index descents can be reduced nearly by half.
This for TODO and needs to be considered later.